python/requirements.txt
asgiref==3.4.1
certifi==2023.7.22
charset-normalizer==2.0.12
Django==3.2.20
django-cors-headers==3.10.1
idna==3.4
mysqlclient==2.1.1
PyJWT==2.4.0
pytz==2023.3
requests==2.27.1
sqlparse==0.4.4
tencentcloud-sdk-python==3.0.958
typing_extensions==4.1.1
urllib3==1.26.16
